moved _get_usb_location back to devices.py
