+++ unit-tests/py/rspy/device_hub.py
@@ -106,14 +106,8 @@ class device_hub(ABC):
         #
         return result
 
-    def get_usb_and_port_location(self, physical_port, hubs):
-        usb_location = None
+    def get_port_by_location(self, physical_port, usb_location, hubs):
         port = None
-        try:
-            usb_location = _get_usb_location( physical_port )
-        except Exception as e:
-            log.e( 'Failed to get usb location:', e )
-
         try:
             port = self._get_port_by_loc(usb_location, hubs)
         except Exception as e:
@@ -130,6 +124,10 @@ def find_all_hubs(vid):
     """
     Yields all hub port numbers
     """
+    # example for Acroname hub:
+    # 24ff:8013 =
+    #   iManufacturer           Acroname Inc.
+    #   iProduct                USBHub3p-3[A]
     from rspy import lsusb
     hubs = set( lsusb.devices_by_vendor( vid ))
     ports = set()
@@ -188,61 +186,3 @@ def _create_ykush():
         return None
     except:
         return None
-
-
-###############################################################################################
-if 'windows' in platform.system().lower():
-    def _get_usb_location( physical_port ):
-        """
-        Helper method to get Windows USB location from registry
-        """
-        if not physical_port:
-            return None
-        # physical port example:
-        #   \\?\usb#vid_8086&pid_0b07&mi_00#6&8bfcab3&0&0000#{e5323777-f976-4f5b-9b55-b94699c46e44}\global
-        #
-        re_result = re.match( r'.*\\(.*)#vid_(.*)&pid_(.*)(?:&mi_(.*))?#(.*)#', physical_port, flags = re.IGNORECASE )
-        dev_type = re_result.group(1)
-        vid = re_result.group(2)
-        pid = re_result.group(3)
-        mi = re_result.group(4)
-        unique_identifier = re_result.group(5)
-        #
-        import winreg
-        if mi:
-            registry_path = "SYSTEM\CurrentControlSet\Enum\{}\VID_{}&PID_{}&MI_{}\{}".format(
-                dev_type, vid, pid, mi, unique_identifier
-                )
-        else:
-            registry_path = "SYSTEM\CurrentControlSet\Enum\{}\VID_{}&PID_{}\{}".format(
-                dev_type, vid, pid, unique_identifier
-                )
-        try:
-            reg_key = winreg.OpenKey( winreg.HKEY_LOCAL_MACHINE, registry_path )
-        except FileNotFoundError:
-            log.e( 'Could not find registry key for port:', registry_path )
-            log.e( '    usb location:', physical_port )
-            return None
-        result = winreg.QueryValueEx( reg_key, "LocationInformation" )
-        # location example: 0000.0014.0000.016.003.004.003.000.000
-        # and, for T265: Port_#0002.Hub_#0006
-        return result[0]
-else:
-    def _get_usb_location( physical_port ):
-        """
-        """
-        if not physical_port:
-            return None
-        # physical port example:
-        # u'/sys/devices/pci0000:00/0000:00:14.0/usb2/2-3/2-3.3/2-3.3.1/2-3.3.1:1.0/video4linux/video0'
-        #
-        split_location = physical_port.split( '/' )
-        if len(split_location) > 4:
-            port_location = split_location[-4]
-        elif len(split_location) == 1:
-            # Recovery devices may have only the relevant port: e.g., L515 Recovery has "2-2.4.4-84"
-            port_location = physical_port
-        else:
-            raise RuntimeError( f"invalid physical port '{physical_port}'" )
-        # location example: 2-3.3.1
-        return port_location
